+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [2.0.0]
  - [BREAKING] `Parsers` have been replaced with `Marshallers`, allowing bi-directional conversions
+ - Added `EightBall.marshall` as a way to output the Feature list to an external format (e.g. to create a JSON file)
  - Added `EightBall.features` as a shortcut to `EightBall.provider.features`
  - Testing framework has been moved from Minitest to rspec
  - Updated dev dependencies